+++ apps/server/src/trpc/index.ts
@@ -5,11 +5,13 @@ import { search } from './search'
 import { degree } from './degree'
 import { getOne } from './get-one'
 import { list } from './list'
+import { login } from './login'
 
 export const appRouter = createRouter()
   /** root-level stuff */
   .merge(getOne)
   .merge(list)
+  .merge(login)
   /** one-up from root */
   .merge('search/', search)
   .merge('degree/', degree)
