+++ apps/server/src/trpc/user.ts
@@ -241,34 +241,6 @@ export const user = createRouter()
   })
 
   /**
-   * user login
-   */
-  .mutation('login', {
-    meta: {
-      openapi: {
-        enabled: true,
-        method: 'POST',
-        path: '/user/login',
-        tags: ['User'],
-        summary: 'User login',
-        description:
-          'Performs user login. If there is no associated user in the database, creates a user with an empty degree and empty graph.',
-      },
-    },
-    input: z.object({
-      provider: z.string(),
-      providerId: z.string().min(1),
-      email: z.string().email(),
-    }),
-    output: entities.User,
-    async resolve({ input }) {
-      return api
-        .userLogin(input.email, input.provider, input.providerId)
-        .then(flatten.user)
-    },
-  })
-
-  /**
    * get a user by email
    */
   .query('get-by-email', {
